+++ crates/biome_js_semantic/src/semantic_model/scope.rs
@@ -14,7 +14,7 @@ pub(crate) struct SemanticModelScopeData {
     pub(crate) children: Vec<usize>,
     // All bindings of this scope (points to SemanticModelData::bindings)
     pub(crate) bindings: Vec<usize>,
-    // Map pointing to the [bindings] vec  of each bindings by its name
+    // Map pointing to the [bindings] vec of each bindings by its name
     pub(crate) bindings_by_name: FxHashMap<TokenText, usize>,
     // All read references of a scope
     pub(crate) read_references: Vec<SemanticModelScopeReference>,